+++ examples/file-based-routing/packages/client/src/App.tsx
@@ -1,6 +1,9 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React, { useMemo } from 'react';
-import { api } from './generated/react-query-client';
+import {
+  ApiClientProvider,
+  createApiClient,
+} from './generated/react-query-client';
 import { Shop } from './shop/Shop';
 
 const API_URL = 'http://localhost:3000';
@@ -8,15 +11,15 @@ const API_URL = 'http://localhost:3000';
 export function App() {
   const queryClient = useMemo(() => new QueryClient(), []);
   const apiClient = useMemo(
-    () => api.createReactQueryClient({ baseURL: API_URL }),
+    () => createApiClient({ baseURL: API_URL }),
     [],
   );
 
   return (
     <QueryClientProvider client={queryClient}>
-      <api.Provider client={apiClient}>
+      <ApiClientProvider client={apiClient}>
         <Shop />
-      </api.Provider>
+      </ApiClientProvider>
     </QueryClientProvider>
   );
 }