+++ packages/test/test-react-query-client/src/App.tsx
@@ -1,6 +1,9 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React, { useMemo } from 'react';
-import { api } from '../../test-app-express-cjs/generated/react-query-client';
+import {
+  ApiClientProvider,
+  createApiClient,
+} from '../../test-app-express-cjs/generated/react-query-client';
 import { UserPage } from './UserPage';
 
 interface AppProps {
@@ -11,7 +14,7 @@ export function App({ children }: AppProps = {}) {
   const queryClient = useMemo(() => new QueryClient({}), []);
   const apiClient = useMemo(
     () =>
-      api.createReactQueryClient({
+      createApiClient({
         baseURL: `http://localhost:${import.meta.env.VITE_WEB_API_PORT}`,
       }),
     [],
@@ -19,9 +22,9 @@ export function App({ children }: AppProps = {}) {
 
   return (
     <QueryClientProvider client={queryClient}>
-      <api.Provider client={apiClient}>
+      <ApiClientProvider client={apiClient}>
         {children ?? <UserPage />}
-      </api.Provider>
+      </ApiClientProvider>
     </QueryClientProvider>
   );
 }